+++ test/instagram.test.ts
@@ -1,11 +1,11 @@
 import { test, expect } from "vitest"
-import { getFollowerCount, getIgSessionId } from "../src"
+import { getFollowerCount, getIgSessionId, isAxiosError } from "../src"
+
+const IG_USER = import.meta.env.VITE_IG_USER
+const IG_PASSWORD = import.meta.env.VITE_IG_PASSWORD
 
 test("instagram", async () => {
-  const sessionId = await getIgSessionId(
-    "dummy_ig_user_egoist",
-    import.meta.env.VITE_IG_PASSWORD,
-  )
+  const sessionId = await getIgSessionId(IG_USER, IG_PASSWORD)
   const count = await getFollowerCount({
     type: "instagram",
     username: "hello",
@@ -18,16 +18,18 @@ test("instagram", async () => {
 
 test("instagram user not found", async () => {
   try {
-    const sessionId = await getIgSessionId(
-      "dummy_ig_user_egoist",
-      import.meta.env.VITE_IG_PASSWORD,
-    )
+    const sessionId = await getIgSessionId(IG_USER, IG_PASSWORD)
     await getFollowerCount({
       type: "instagram",
       username: "hello_xx88aa",
       sessionId,
     })
+    expect.fail()
   } catch (error: any) {
-    expect(error.message).toContain("user not found")
+    if (isAxiosError(error)) {
+      expect(error.response?.status).toBe(404)
+    } else {
+      expect.fail()
+    }
   }
 })